Fix random block issues (#261)
* Possible solution for incorrect random value calculation.

To get next block it calculated random value 2 times. This increases a chance for a rare block to spawn.
Also, this change separates block and chest randomness, just to avoid influencing block spawning chances.

* The arraylist is not suited for constantly removing first element and adding one at the end.
The Queue is exactly that does it. And linked list just works in this case better.