Release 1.11.0 (#285)
* Fix random block issues (#261)

* Possible solution for incorrect random value calculation.

To get next block it calculated random value 2 times. This increases a chance for a rare block to spawn.
Also, this change separates block and chest randomness, just to avoid influencing block spawning chances.

* The arraylist is not suited for constantly removing first element and adding one at the end.
The Queue is exactly that does it. And linked list just works in this case better.

* Do not replace water with water.

This will replace with water-only blocks that are not water, similar to the water-logged blocks.
People complained that flowing water was not accepted as a valid block for water animals.

Fixes #265

* 62.5% and 75% had identical filling of the scale. I fixed it now everything should be right. (#277)

* ADDED WARDEN SOUND(EVENT mob-warning) (#281)

I added a line that produces the sound of the WARDEN spawn and also the color combination of it

* Update to Spigot 1.19

* Custom Block support (#283)

* setCustomBlock

* Config & OneBlockCustomBlockCreator

* revert reformat

* just custom-blocks

* use Blocks with new approach

* simply return instead of default value

* Migrate to BentoBox API 1.21

This migration allows to define any custom flag names for default settings. If the flag does not exist, it will just ignore it and not remove.

* Remove unused methods.

* Fixes #276 and #278

Co-authored-by: Oh hi Mark! <61845375+MrMarL@users.noreply.github.com>
Co-authored-by: Marmur2020 <109208530+Marmur2020@users.noreply.github.com>
Co-authored-by: Huynh Tien <huynhqtienvtag@gmail.com>