Fixes a bug where non-AOneBlock islands were stored
Commit a507029 introduced a bug where every island operated by the
server was loaded into the AOneBlock cache and a Hologram placed at the
center location of every island on the server even if the hologram was
blank. In writing the top ten addon for AOneBlock these erroneous
islands were discovered. This commit stops this from happening and
cleans up the database of all the unonwed and non-AOneBlock islands.