Fix for null players, i.e., minions at end of phase.
If a minion broke the block at the end of the phase, the NPEs would
occur as the newer code to check requirements and do placeholders was
assuming a real player. This change protects that. Note that minions
cannot be used to progress past a phase if there are any restrictions on
the level change.