Change how holograms are placed and removed #436
Hologram entities do not remain loaded even if marked as persistent.
Further, if the chunks are not loaded, they cannot be referenced by
UUID. So, in the end if I have to load the chunk where the entity is, I
may as well just store the location of the hologram. Even that isn't
required because the hologram is always in the same place relative to
the island, so the cache is only required to clear holograms on a server
shutdown.